[오류수정] feat: implement core project structure with data sources and repositories

+++ domain/src/main/java/com/example/domain/model/base/ProjectChannel.kt
@@ -1,15 +1,13 @@
 package com.example.domain.model.base
 
-import com.example.domain.model._new.enum.ProjectChannelType
-import com.google.firebase.Timestamp
+import com.example.domain.model.enum.ProjectChannelType
 import com.google.firebase.firestore.DocumentId
-import com.google.firebase.firestore.ServerTimestamp
 import java.time.Instant
 
 data class ProjectChannel(
     @DocumentId val id: String = "",
     val channelName: String = "",
-    val order: Int = 0,
+    val order: Double = 0.0,
     val channelType: ProjectChannelType = ProjectChannelType.MESSAGES, // "MESSAGES", "TASKS" 등
     val createdAt: Instant? = null,
     val updatedAt: Instant? = null
